Replace (some) build paths in documentation.
This should take care of instances where we called `code` or are printing
the net of a function.  There are likely more yet to do!

Right now, this is too Debian-specific (assuming /usr/share), so this
would need modification before going upstream.

Also got rid of the sed band-aid in d/rules, which wasn't hitting the
.db docs in /usr/lib and missed line breaks in the middle of paths.