Remove C-m from M2-mode-map
It is currently bound to `M2-newline-and-indent`, which calls
`newline` and then indents.  However, `newline` by itself already
indents the next line (provided that the user has
`electric-indent-mode` enabled, which is the default), and so calling
this specialized function is unnecessary.  Furthermore, C-m is already
bound to `newline` in `global-map`, so there's no reason to bind it
again.

Also, if the user has turned off `electric-indent-mode`, then C-m will
have the desired behavior -- no electric indents.