M2-electric-tab modifications
We now always tab electrically, even when not at the beginning of a
line or on a blank line.  This has several advantages:

* The function is now idempotent.
* The "newline" function (when electric-indent-mode is on, the
  default) will do the correct thing, so we can deprecate
  "M2-newline-and-indent".
* The new behavior is consistent with other common major modes.

We also always delete leading whitespace and re-indent to clean up
tabs and spaces.

Note that it is still possible to insert a tab character with C-q TAB.