Print entire example/test error output in verbose mode
Previously, when installPackage() or check() was called with Verbose =>
true and an error occurred, we were only shown the last 10 lines of
the log.  This may not be sufficient, especially in the case of
segmentation faults with long stacktraces.

Instead, we use a regular expression to determine where the error
occurred, print the 9 previous lines (mimicking the previous
behavior), and then everything afterwards.

If the regular expression doesn't match anything, then we fall back on
the previous behavior and use tail.