Use RTLD_DEFAULT in dlsym to find symbols w/o library handle
Previously, we hardcoded it as (void *)0, but this is only the case in
Linux (it's (void *)-2 on macOS/BSD).

We add "use system" to the beginning since that will add a #define for
the _GNU_SOURCE macro to ffi-exports.h (since that's in system.d's
"declarations").  We need this to be defined before dlfcn.h is
included, or RTLD_DEFAULT won't be defined.  And dlfcn.h is actually
being included before we explicitly include it in "headers" via a GC header.