Use "dereference" for getting array/struct/union values
Also for casting a foreign object to a different type.

By just using "T ptr", we were allocating a new pointer pointing to
ptr when T was a ForeignPointerType.  But we just want to use the
existing pointer, so use "dereference_T ptr", just as in the return
values of foreign functions.