Don't flag timestamps as file locations
Our regular expressions flag any string ending in ":[0-9]+:[0-9]+" as
a file location, including timestamps, e.g. the one in
version#"compile time".

We add a new element to compilation-transform-file-match-alist so that
compilation mode knows not to make any filenames consisting of two
digits (i.e., the hour, since the minute and seconds are interpreted
as row and column number) clickable.