Add non-Groebner basis multivariate polynomial division algorithm
We add a new quotientRemainder(RingElement, List) method for dividing
a polynomial by a list of polynomials without using Groebner bases.
In particular, it uses the algorithm presented by Cox, Little, and
O'Shea in Ideals, Varieties, and Algorithms, Section 2.3, Theorem 3.
The quotient and remainder will depend on the order of the elements of
the list of divisors.