Update autotools flint build to use submodule
Note that VERSION is now only used for naming the build directory.  We
update it to match the current head of the flint submodule, but it
doesn't really matter what it is.

Also run bootstrap.sh to generate configure script (not present in
git) and drop the deprecated "--disable-tls" configure option.