Set random seed during Cremona tests
Previously, calling random(QQ) called two GMP random functions, but
now it only calls one MPFR one.  So now we are earlier in the random
number sequence when we get to the next line, which was causing the
assertion to fail.  So we reset the random seed to 1, after which it
will pass.