Remove JSONEncoder class
We can deal with everything by just using options to the toJSON
method.  In particular, we add an IndentLevel option which keeps track
of the current indentation level, which was the primary purpose of the
old JSONEncoder class.

We also change some of the internal methods to functions.  There was
no multiple dispatch, so using methods was unnecessary.